+++ src/main/java/com/hit/devicemanage/entity/Devicegroup.java
@@ -6,12 +6,12 @@ import jakarta.persistence.*;
 public class Devicegroup {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long gid;
+    private Integer gid;
 
     private String gname;
     private String gcode;
 
-    public Devicegroup(Long gid, String gname, String gcode) {
+    public Devicegroup(Integer gid, String gname, String gcode) {
         this.gid = gid;
         this.gname = gname;
         this.gcode = gcode;
@@ -21,10 +21,10 @@ public class Devicegroup {
 
     }
 
-    public Long getGid() {
+    public Integer getGid() {
         return gid;
     }
-    public void setGid(Long gid) {
+    public void setGid(Integer gid) {
         this.gid = gid;
     }
     public String getGname() {
