+++ src/main/java/com/hit/devicemanage/service/DevicegroupRepository.java
@@ -3,6 +3,6 @@ package com.hit.devicemanage.service;
 import com.hit.devicemanage.entity.Devicegroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DevicegroupRepository  extends JpaRepository<Devicegroup,Long>{
+public interface DevicegroupRepository  extends JpaRepository<Devicegroup,Integer>{
     public Devicegroup findByGcode(String gcode);
 }
