+++ src/main/java/com/hit/devicemanage/service/SiteuserRepository.java
@@ -3,7 +3,7 @@ package com.hit.devicemanage.service;
 import com.hit.devicemanage.entity.Siteuser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SiteuserRepository extends JpaRepository<Siteuser, Long> {
+public interface SiteuserRepository extends JpaRepository<Siteuser, Integer> {
 
     Siteuser findByUname(String uname);
